Make ArgumentFilter and ArgumentMap private static classes
?
